Spätzle: moved more functions
